Delete malloc of buffer, add free tp finder
